Parial fix for #23.
Adds multiple damage types to the damage types dropdown for weapon damage. I didn't create a new entry for each possible combination, instead I went the AoN and checked to see what combinations currently exist, and just added those. The next step is to pass this info on to the roll chat card and then do the correct calculations to make sure that weapons with multiple damage types are spliting the damage evenly between the two types. I aslo added damage types where they can be one OR the other as well.