Continued work on modifiers.
I've got most of the backend pieces in place, I just need to finish the front end. I still need to do modifier calculations for a few things, but I think it's good enough for a initial release of the system. I just need to test to make sure everything is calculating correctly.