"pseudo-gunnery" implementation
Implemented the "pseudo-gunnery skill" option. When gunnery is added to an NPC crew, ranks will be uneditable and show a "-". Changed gunner actions to use the max of BAB/Pilot ranks/Gunnery Mod. Gunnery mod returns as 0 from PCs, so this will choose the correct modifier based on if it's a PC or NPC gunner.