Allow SFRPG rolls without chat messages (#845)
* chatMessage changes

I have a module that would like to roll ability, skills and saving throws.  Problem is I don't want the request to automatically open a Chat Message.  I've attempted some interested code to work around it, but DnD5e has a handy option of requesting that a dice roll doesn't open a ChatMessage, and I'd very kindly like to add it to Starfinder.

* chatMessage changes

* chatMessage changes

* chatMessage changes

* removed unecessary condition

* removed unecessary condition

Co-authored-by: IronMonk88 <twilight.lumineer@gmail.com>