Fixing issues with Infinite Loops and incorrect usage
Okay, so I found a huge bug in the Starfinder code that reared its ugly head when I tested it recently. The problem is that either in the newest version of Foundry, or its just been hidden for a while, is an infinite loop. What was happening is the calculation for the modifiers was calling... a calculation for the modifiers. The entire chain was repeating until a StackOverflowException happened, and that exception was logged, discarded, and the system continued as normal. I fixed it by removing the infinite loop in the various calculate js files. I also fixed my own code when it came down to Hide Untrained. It works now.