Fix Infinite Recursion Bugs
This fixes a huge Recursion issue in the Rules section. All of the calculate- js are being called by the stackModifiers.process system. However, inside of these functions stackModifiers.process was being called. The reason why it didn't result in a StackOverflow was because in calculate-encumbrance.js there was an error that stopped the stack originally. Fixing it resulted in showing the StackOverflowException (since it was going into infinite recursion).