Fixes #1233
The extra regex ^ was causing the expression to treat the first character as a splitter, but, with how JavaScript Regex Splitters work, it will see that ^ is index 0, and then use that character at index 0 as the split, so if you have a string:

"so"
A regex with ([^,;])+ will create three groups:
"","o",""
This is because it finds that index 0 is a split character, and creates an empty string for the first item, and then puts index 1 as the second item. The third item is something something extra that RegEx likes to do. I'm sure I could look it up, but this just fixes the issue entirely.