Repaired "picking up dropped items" (#1314)
* Repaired "picking up dropped items"

If a dropped item is clicked on and pulled into the inventory, it results in a "processedDroppedData", transfering the instance of the dropped object under "parsedDragData.items[0]". The item is not given as ItemCollection but as an object already, and by casting "toObject()" it simply breaks at this point. As such things can't be taken into the inventory. Removing that wrong command makes it work as intendet.

* Update src/module/actor/sheet/base.js

---------

Co-authored-by: LebombJames <77904738+LebombJames@users.noreply.github.com>