Pivot `buildRoll()` to return data rather than consume a callback.
Converts `RollTree.prototype.buildRoll` to return data rather than take
a callback as a parameter. The hope is that this will make data flow
more explicit and therefore easier to reason about. The diff looks less
impressive if you ignore whitespace changes.

This change helped with identifying and unifying duplicate behavior in
`DiceSFRPG.createRoll`. At some point in development the duty of
`dialogOptions.skipUI` was taken on by `RollTree.prototype.buildRoll`,
so `createRoll()` can be relieved of the responsibility.