Convert `RollTree.prototype.buildRoll` to static, add sync vesion
Instances of `RollTree` are only used by the roll dialog, and their
usefulness is largely over after `buildRoll` is called. In theory they
could have been reused, but in practice they are so stateful that new
`RollTree` objects were created for each invocation of `buildRoll`.
Converting `buildRoll` to a static method that creates a local
`RollTree` object captures how the class was being used in practice.

Because only showing the UI is asynchronous, the rest of `buildRoll`
could be factored into a synchronous function, which aided the creation
of `buildRollSync`. `buildRollSync` is useful exactly once, but it's in
the rules engine code, which was not itself async. It had instead been
relying on the coincidence that the callback passed to `buildRoll` would
be called synchronously if showing the UI were skipped. When `buildRoll`
was refactored to return data rather than consume a callback, this
behavior was broken, but now in this commit restored.

Along side this change `RollTree.prototype.populate` was changed to take
the context as a parameter, and the context is no longer stored on the
tree. The intent of this was to make the relationship between `populate`
and the `RollContext` more explicit, instead of by implication, when
reading `RollDialog.prototype._onSelectorChanged`, which changes the
context then re-populates the stored `RollTree` object.