Roll Dialog: do not modify `formula`, add `simplifiedFormula`
Invert the existing semantics, where `RollDialog.prototype.getData`
simplifies formulae in-place, then adds an `originalFormula` field that
can be presented to the user. Now `getData` will leave the original
`formula` (or `modifier` for modifiers) field alone and add a
`simplifiedFormula` (or `simplifiedModifier`) field that the template
will use.

For modifiers, this has no knock-on effects, as `getData` clones
the available modifiers, but for damage groups / parts this will have an
effect. Damage parts were not being cloned, so modifications to
`formula` would be visible to callers.