Feature Macro Triggers
Adds a Feature Macro Trigger section to features' Usage & Action
sections. If the UUID provided exists and can be executed, then it will
be. Otherwise it's silently ignored. These macros fire after the feat
activates or deactivates. Script macros can access the feat's owner as
`actor` and the feat itself as `scope.item`.

Dropping anything onto the input field will attempt to copy it's UUID
into the input field, and if that UUID exists a little badge previewing
the item will show up. I ripped this idea directly from what I've seen
the PF2e system do. Clicking the badge will open the item's sheet. If
the user cannot execute that item, such as if it's not a macro, the name
will appear red to indicate an issue.